chore: add register to user handler and fix code
